+++ github/__init__.py
@@ -1,4 +1,19 @@
 from .branches import GithubBranch
 from .branches import GithubBranchApi
+from .packages import ContainerPackage
+from .packages import GithubContainerRegistryOrgApi
+from .packages import GithubContainerRegistryUserApi
+from .packages import create_registry_api
+from .pullrequest import GithubPullRequestApi
+from .ratelimit import GithubRateLimitApi
 
-__all__ = ["GithubBranch", "GithubBranchApi"]
+__all__ = [
+    "ContainerPackage",
+    "GithubBranch",
+    "GithubBranchApi",
+    "GithubContainerRegistryOrgApi",
+    "GithubContainerRegistryUserApi",
+    "GithubPullRequestApi",
+    "GithubRateLimitApi",
+    "create_registry_api",
+]